+++ packages/gatsby-plugin-image/src/index.ts
@@ -1,0 +1,4 @@
+export { GatsbyImage } from "./components/gatsby-image.server"
+export { Placeholder } from "./components/placeholder"
+export { MainImage } from "./components/main-image"
+export { StaticImage } from "./components/static-image.server"