+++ packages/gatsby/src/bootstrap/load-plugins/__tests__/load-plugins.ts
@@ -1,6 +1,24 @@
+jest.mock(`gatsby-cli/lib/reporter`, () => {
+  return {
+    error: jest.fn(),
+    panic: jest.fn(),
+    log: jest.fn(),
+    warn: jest.fn(),
+    success: jest.fn(),
+    info: jest.fn(),
+  }
+})
+const mockProcessExit = jest.spyOn(process, `exit`).mockImplementation(() => {})
 import { loadPlugins } from "../index"
 import { slash } from "gatsby-core-utils"
+import reporter from "gatsby-cli/lib/reporter"
 import { IFlattenedPlugin } from "../types"
+
+afterEach(() => {
+  Object.keys(reporter).forEach(method => {
+    reporter[method].mockClear()
+  })
+})
 
 describe(`Load plugins`, () => {
   /**
@@ -108,20 +126,7 @@ describe(`Load plugins`, () => {
       expect(plugins).toEqual(
         expect.arrayContaining([
           expect.objectContaining({
-            browserAPIs: [],
-            id: ``,
             name: `gatsby-plugin-typescript`,
-            nodeAPIs: [
-              `resolvableExtensions`,
-              `onCreateBabelConfig`,
-              `onCreateWebpackConfig`,
-            ],
-            pluginOptions: {
-              plugins: [],
-            },
-            resolve: ``,
-            ssrAPIs: [],
-            version: `1.0.0`,
           }),
         ])
       )
@@ -150,13 +155,16 @@ describe(`Load plugins`, () => {
             id: ``,
             name: `gatsby-plugin-typescript`,
             nodeAPIs: [
+              `pluginOptionsSchema`,
               `resolvableExtensions`,
               `onCreateBabelConfig`,
               `onCreateWebpackConfig`,
             ],
             pluginOptions: {
-              plugins: [],
+              allExtensions: false,
+              isTSX: false,
               jsxPragma: `h`,
+              plugins: [],
             },
             resolve: ``,
             ssrAPIs: [],
@@ -185,6 +193,173 @@ describe(`Load plugins`, () => {
       // TODO: I think we should probably be de-duping, so this should be 1.
       // But this test is mostly here to ensure we don't add an _additional_ gatsby-plugin-typescript
       expect(tsplugins.length).toEqual(2)
+    })
+  })
+
+  describe(`plugin options validation`, () => {
+    it(`throws a structured error with invalid plugin options`, async () => {
+      const invalidPlugins = [
+        {
+          resolve: `gatsby-plugin-google-analytics`,
+          options: {
+            trackingId: 123,
+            anonymize: `not a boolean`,
+          },
+        },
+        {
+          resolve: `gatsby-plugin-google-analytics`,
+          options: {
+            anonymize: `still not a boolean`,
+          },
+        },
+      ]
+      await loadPlugins({
+        plugins: invalidPlugins,
+      })
+
+      expect(reporter.error as jest.Mock).toHaveBeenCalledTimes(
+        invalidPlugins.length
+      )
+      expect((reporter.error as jest.Mock).mock.calls[0])
+        .toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "context": Object {
+              "pluginName": "gatsby-plugin-google-analytics",
+              "validationErrors": Array [
+                Object {
+                  "context": Object {
+                    "key": "trackingId",
+                    "label": "trackingId",
+                    "value": 123,
+                  },
+                  "message": "\\"trackingId\\" must be a string",
+                  "path": Array [
+                    "trackingId",
+                  ],
+                  "type": "string.base",
+                },
+                Object {
+                  "context": Object {
+                    "key": "anonymize",
+                    "label": "anonymize",
+                    "value": "not a boolean",
+                  },
+                  "message": "\\"anonymize\\" must be a boolean",
+                  "path": Array [
+                    "anonymize",
+                  ],
+                  "type": "boolean.base",
+                },
+              ],
+            },
+            "id": "11331",
+          },
+        ]
+      `)
+      expect((reporter.error as jest.Mock).mock.calls[1])
+        .toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "context": Object {
+              "pluginName": "gatsby-plugin-google-analytics",
+              "validationErrors": Array [
+                Object {
+                  "context": Object {
+                    "key": "anonymize",
+                    "label": "anonymize",
+                    "value": "still not a boolean",
+                  },
+                  "message": "\\"anonymize\\" must be a boolean",
+                  "path": Array [
+                    "anonymize",
+                  ],
+                  "type": "boolean.base",
+                },
+              ],
+            },
+            "id": "11331",
+          },
+        ]
+      `)
+      expect(mockProcessExit).toHaveBeenCalledWith(1)
+    })
+
+    it(`defaults plugin options to the ones defined in the schema`, async () => {
+      let plugins = await loadPlugins({
+        plugins: [
+          {
+            resolve: `gatsby-plugin-google-analytics`,
+            options: {
+              trackingId: `fake`,
+            },
+          },
+        ],
+      })
+
+      plugins = replaceFieldsThatCanVary(plugins)
+
+      expect(
+        plugins.find(plugin => plugin.name === `gatsby-plugin-google-analytics`)
+          .pluginOptions
+      ).toEqual({
+        // All the options that have defaults are defined
+        anonymize: false,
+        exclude: [],
+        head: false,
+        pageTransitionDelay: 0,
+        plugins: [],
+        respectDNT: false,
+        trackingId: `fake`,
+      })
+    })
+
+    it(`validates subplugin schemas`, async () => {
+      await loadPlugins({
+        plugins: [
+          {
+            resolve: `gatsby-transformer-remark`,
+            options: {
+              plugins: [
+                {
+                  resolve: `gatsby-remark-autolink-headers`,
+                  options: {
+                    maintainCase: `should be boolean`,
+                  },
+                },
+              ],
+            },
+          },
+        ],
+      })
+
+      expect(reporter.error as jest.Mock).toHaveBeenCalledTimes(1)
+      expect((reporter.error as jest.Mock).mock.calls[0])
+        .toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "context": Object {
+              "pluginName": "gatsby-remark-autolink-headers",
+              "validationErrors": Array [
+                Object {
+                  "context": Object {
+                    "key": "maintainCase",
+                    "label": "maintainCase",
+                    "value": "should be boolean",
+                  },
+                  "message": "\\"maintainCase\\" must be a boolean",
+                  "path": Array [
+                    "maintainCase",
+                  ],
+                  "type": "boolean.base",
+                },
+              ],
+            },
+            "id": "11331",
+          },
+        ]
+      `)
+      expect(mockProcessExit).toHaveBeenCalledWith(1)
     })
   })
 })