+++ packages/gatsby/src/services/initialize.ts
@@ -1,7 +1,7 @@
 import _ from "lodash"
 import { slash } from "gatsby-core-utils"
 import fs from "fs-extra"
-import md5File from "md5-file/promise"
+import md5File from "md5-file"
 import crypto from "crypto"
 import del from "del"
 import path from "path"
@@ -226,15 +226,11 @@ export async function initialize({
   // The last, gatsby-node.js, is important as many gatsby sites put important
   // logic in there e.g. generating slugs for custom pages.
   const pluginVersions = flattenedPlugins.map(p => p.version)
-  const hashes = await Promise.all([
+  const hashes: any = await Promise.all([
     !!process.env.GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES,
     md5File(`package.json`),
-    Promise.resolve(
-      md5File(`${program.directory}/gatsby-config.js`).catch(() => {})
-    ), // ignore as this file isn't required),
-    Promise.resolve(
-      md5File(`${program.directory}/gatsby-node.js`).catch(() => {})
-    ), // ignore as this file isn't required),
+    md5File(`${program.directory}/gatsby-config.js`).catch(() => {}), // ignore as this file isn't required),
+    md5File(`${program.directory}/gatsby-node.js`).catch(() => {}), // ignore as this file isn't required),
   ])
   const pluginsHash = crypto
     .createHash(`md5`)
@@ -255,7 +251,22 @@ export async function initialize({
     `)
   }
   const cacheDirectory = `${program.directory}/.cache`
-  if (!oldPluginsHash || pluginsHash !== oldPluginsHash) {
+  const publicDirectory = `${program.directory}/public`
+
+  // .cache directory exists in develop at this point
+  // so checking for .cache/json as a heuristic (could be any expected file)
+  const cacheIsCorrupt =
+    fs.existsSync(`${cacheDirectory}/json`) && !fs.existsSync(publicDirectory)
+
+  if (cacheIsCorrupt) {
+    reporter.info(reporter.stripIndent`
+      We've detected that the Gatsby cache is incomplete (the .cache directory exists
+      but the public directory does not). As a precaution, we're deleting your site's
+      cache to ensure there's no stale data.
+    `)
+  }
+
+  if (!oldPluginsHash || pluginsHash !== oldPluginsHash || cacheIsCorrupt) {
     try {
       // Attempt to empty dir if remove fails,
       // like when directory is mount point
@@ -290,7 +301,7 @@ export async function initialize({
   await fs.ensureDir(cacheDirectory)
 
   // Ensure the public/static directory
-  await fs.ensureDir(`${program.directory}/public/static`)
+  await fs.ensureDir(`${publicDirectory}/static`)
 
   activity.end()