+++ packages/gatsby-image/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.20...gatsby-image@2.4.21) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-image
+
 ## [2.4.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-image@2.4.19...gatsby-image@2.4.20) (2020-09-28)
 
 **Note:** Version bump only for package gatsby-image