+++ packages/gatsby-plugin-postcss/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@3.0.3...gatsby-plugin-postcss@3.0.4) (2020-10-26)
+
+### Bug Fixes
+
+- **deps:** update minor and patch for gatsby-plugin-postcss ([#27280](https://github.com/gatsbyjs/gatsby/issues/27280)) ([ed75748](https://github.com/gatsbyjs/gatsby/commit/ed75748805eb1fa10e2446388b5992a51133f182))
+
+## [3.0.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@3.0.2...gatsby-plugin-postcss@3.0.3) (2020-10-14)
+
+### Bug Fixes
+
+- **gatsby-plugin-postcss:** Delete invalid plugins property from outer options object. ([#27430](https://github.com/gatsbyjs/gatsby/issues/27430)) ([2fd07f9](https://github.com/gatsbyjs/gatsby/commit/2fd07f925113b4e1f578480bf68c93282a7e3dca))
+
+## [3.0.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@3.0.1...gatsby-plugin-postcss@3.0.2) (2020-10-13)
+
+### Bug Fixes
+
+- **gatsby-plugin-postcss:** Add required additional level of nesting for postcss-loader options. ([#27418](https://github.com/gatsbyjs/gatsby/issues/27418)) ([8756441](https://github.com/gatsbyjs/gatsby/commit/8756441819138e12c14460c8a4e5c3c28fe338b9))
+
+## [3.0.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@3.0.0...gatsby-plugin-postcss@3.0.1) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-plugin-postcss
+
 # [3.0.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-postcss@2.3.13...gatsby-plugin-postcss@3.0.0) (2020-10-01)
 
 **Note:** Version bump only for package gatsby-plugin-postcss