+++ packages/gatsby-theme/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme@0.1.5...gatsby-theme@0.1.6) (2020-10-08)
+
+**Note:** Version bump only for package gatsby-theme
+
 ## [0.1.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme@0.1.4...gatsby-theme@0.1.5) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-theme